+++ yellowstone-grpc-geyser/src/plugin.rs
@@ -234,7 +234,10 @@ impl GeyserPlugin for Plugin {
                 ReplicaTransactionInfoVersions::V0_0_1(_info) => {
                     unreachable!("ReplicaAccountInfoVersions::V0_0_1 is not supported")
                 }
-                ReplicaTransactionInfoVersions::V0_0_2(info) => info,
+                ReplicaTransactionInfoVersions::V0_0_2(_info) => {
+                    unreachable!("ReplicaAccountInfoVersions::V0_0_2 is not supported")
+                }
+                ReplicaTransactionInfoVersions::V0_0_3(info) => info,
             };
 
             let message = Message::Transaction(MessageTransaction::from_geyser(transaction, slot));