+++ yellowstone-grpc-proto/src/lib.rs
@@ -46,13 +46,12 @@ pub mod convert_to {
         super::prelude as proto,
         solana_clock::UnixTimestamp,
         solana_message::{
-            compiled_instruction::CompiledInstruction,
-            v0::{LoadedMessage, MessageAddressTableLookup},
-            LegacyMessage, MessageHeader, SanitizedMessage,
+            compiled_instruction::CompiledInstruction, v0::MessageAddressTableLookup,
+            MessageHeader, VersionedMessage,
         },
         solana_pubkey::Pubkey,
         solana_signature::Signature,
-        solana_transaction::sanitized::SanitizedTransaction,
+        solana_transaction::versioned::VersionedTransaction,
         solana_transaction_context::TransactionReturnData,
         solana_transaction_error::TransactionError,
         solana_transaction_status::{
@@ -61,20 +60,20 @@ pub mod convert_to {
         },
     };
 
-    pub fn create_transaction(tx: &SanitizedTransaction) -> proto::Transaction {
+    pub fn create_transaction(tx: &VersionedTransaction) -> proto::Transaction {
         proto::Transaction {
             signatures: tx
-                .signatures()
+                .signatures
                 .iter()
                 .map(|signature| <Signature as AsRef<[u8]>>::as_ref(signature).into())
                 .collect(),
-            message: Some(create_message(tx.message())),
+            message: Some(create_message(&tx.message)),
         }
     }
 
-    pub fn create_message(message: &SanitizedMessage) -> proto::Message {
+    pub fn create_message(message: &VersionedMessage) -> proto::Message {
         match message {
-            SanitizedMessage::Legacy(LegacyMessage { message, .. }) => proto::Message {
+            VersionedMessage::Legacy(message) => proto::Message {
                 header: Some(create_header(&message.header)),
                 account_keys: create_pubkeys(&message.account_keys),
                 recent_blockhash: message.recent_blockhash.to_bytes().into(),
@@ -82,7 +81,7 @@ pub mod convert_to {
                 versioned: false,
                 address_table_lookups: vec![],
             },
-            SanitizedMessage::V0(LoadedMessage { message, .. }) => proto::Message {
+            VersionedMessage::V0(message) => proto::Message {
                 header: Some(create_header(&message.header)),
                 account_keys: create_pubkeys(&message.account_keys),
                 recent_blockhash: message.recent_blockhash.to_bytes().into(),